Disable cloud-init from running in upgrade container
The configuration done via cloud-init generally should not be needed for
the upgrade container, and it can actually cause problems if it races
with the upgrade logic. For example, we use cloud-init to configure the
APT sources on our "dev" images, and this can race with the upgrade
logic that also modifies the APT sources (to point to the APT repository
of our unpacked upgrade image). Thus, this change disables all of the
cloud-init related services from running within the upgrade container.