DLPX-70169 cloud-init and related services need override.conf to prevent them from running within upgrade container
This is a continuation of 144e73fc0e32c968f6d1f4ed4669384c1c4763ed.

As described in 144e73f, we don't want the cloud-init related services
to run within an upgrade container. Unfortunately, that prior commit
does not ensure the services don't run, when there's another service
that depend on a cloud-init service. In our case, when running on GCP,
the "google-startup-scripts" services has a depdency on "cloud-final",

    $ systemctl cat google-startup-scripts | grep Wants
    Wants=cloud-final.service

causing the cloud-init service to run even though it should have
otherwise been disabled (due to the changes in 144e73f).

This change adds configuration to all of the cloud-init services such
that they won't run in the upgrade container; even if they're
inadvertently started due to a dependency, as was our case on GCP.