Generate network config on first and subsequent boots
This change modifies when cloud-init will generate the network config
for the system, such that it will now be generated for the first boot
(which it previously did) and all subsequent boots (which is new). This
way, if the system's network hardware changes (e.g. the MAC address
changes), the network configuration for the system will automatically
adapt on the next boot.

Also note that for a DE, this only applies when using the default
cloud-init generated network configuration. Currently, the cloud-init
config is only used prior to the DE administrator performing any manual
network configuration via the API or CLI. After manual network changes
are made, the DE will no longer use the cloud-init network config, at
which point this change will no longer have any affect.