DLPX-69865 IP configuration is lost if a MAC address changes (#21)
To make the network configuration more resilient to changes of the MAC
address, this change modifies the netplan configuration generated by
cloud-init to no longer match on the MAC address. This way, if the MAC
address of a NIC changes (which can happen in VM environments), the
network configuration on the system should still work as before.